Refactor: Extract scrollToCategory function
